GetChildSubReaperSystemCall: use pointer to int, not long
This points to a 32-bit value, which can be corrupted if the following
32-bits are not zero.

+++ include/syscalls/prctl.hxx
@@ -371,7 +371,7 @@ public: // functions
 public: // data
 
 	/// Current child-subreaper setting filled in by the kernel.
-	item::PointerToScalar<long> is_subreaper;
+	item::PointerToScalar<int> is_subreaper;
 
 protected: // functions
